Use host.docker.internal to get ip address of docker host
This addresses an issue raised by Ian Dickinson where the
previous mechanism for determining the host IP was not reliable.

There are some complications around updating the /etc/hosts file
to have a lr-pres-tunnel.epimorphics.net as a DNS name for that IP
address.  Currently this is done by starting the container as root
so it can update /etc/hosts and then switch to user tester for
running the tests.

Have tried doing this with a root owned script with its setuid
bit set but that didn't work though it feels like it should.
Have asked a question about that on the docker hub forum.